requirements.txt
asgiref==3.2.10
bleach==3.1.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-cors-headers==3.5.0
django-filter==2.3.0
django-post-office==3.4.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
Markdown==3.2.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
webencodings==0.5.1
zipp==3.1.0

Scrapy~=2.3.0

itemadapter~=0.1.0
